feat: new UI elements, refactoring, fixes

+++ src/core/assembly/assembler.hpp
@@ -11,7 +11,7 @@ class Assembler {
   public:
     Assembler(const Config &config);
     std::vector<uint8_t> assemble(const std::string &source,
-                                  uint32_t load_address = 0);
+                                  uint64_t load_address = 0);
 
   private:
     const Config &config_;
@@ -22,7 +22,7 @@ class Assembler {
     int get_register_index(const std::string &name) const;
     uint64_t
     parse_operand(const std::string &op,
-                  const std::unordered_map<std::string, uint32_t> &labels);
+                  const std::unordered_map<std::string, uint64_t> &labels);
     OperandType determine_operand_type(const std::string &token);
     std::vector<OperandType> get_expected_signature(const Instruction &inst);
 };
